// src/data/node-types/ratings/types/ApiRatingForCarModelVariantRelationship.ts
import type {ApiRelationshipType} from "../../../types/ApiRelationshipType"
import type {ApiNodeType} from "../../../types/ApiNodeType"
import type {ApiCarModelVariantNode} from "../../car-model-variants/types/ApiCarModelVariantNode"

export type ApiRatingForCarModelVariantRelationship = {
    links: {
        self: string
    }
    data: {
        type: ApiNodeType.RATING
        id: number
        attributes: Record<string, string | number | boolean | null>
        data: {
            relationship_id: number
            relationship_name: ApiRelationshipType.RATING_FOR_CAR_MODEL_VARIANT
            partner_node: {
                node_type: ApiNodeType.CAR_MODEL_VARIANT
                data: ApiCarModelVariantNode['attributes']
            }
            created_at: string
            updated_at: string
        }
    }
}